+++ test/solar_calculator_test.dart
@@ -1,3 +1,4 @@
+import 'package:solar_calculator/src/instant.dart';
 import 'package:solar_calculator/src/solarCalculator.dart';
 import 'package:solar_calculator/src/timespan.dart';
 import 'package:test/test.dart';
@@ -6,9 +7,52 @@ void main() {
   const _kTimesAccuracyInSec = 60;
   const _kAnglesAccuracy = 0.05;
 
-  var refDate = DateTime.utc(2021, 05, 02);
+  group('Barcelona with time zone ', () {
+    final instant = Instant(2021, 05, 09, 01, 48, 00, 02);
 
-  var date = refDate.add(Duration(hours: 13, minutes: 25, seconds: 35));
+    final calc = SolarCalculator(instant, 41.387048, 2.17413425);
+
+    test('Sunrise', () {
+      print(calc.sunriseTime);
+      final sunrise =
+          calc.sunriseTime.difference(instant.beginingOfDay().add(Duration(hours: 06, minutes: 39))).inSeconds.abs() <=
+              _kTimesAccuracyInSec;
+      expect(sunrise, true);
+    });
+
+    test('Sunset', () {
+      final sunset =
+          calc.sunsetTime.difference(instant.beginingOfDay().add(Duration(hours: 20, minutes: 57))).inSeconds.abs() <=
+              _kTimesAccuracyInSec;
+      expect(sunset, true);
+    });
+
+    test('Transit', () {
+      final transit = calc.sunTransitTime
+              .difference(instant.beginingOfDay().add(Duration(hours: 13, minutes: 47)))
+              .inSeconds
+              .abs() <=
+          _kTimesAccuracyInSec;
+      expect(transit, true);
+    });
+
+    test('Sun Declination', () {
+      final declination = (calc.sunEquatorialPosition.declination - 17.36).abs() <= _kAnglesAccuracy;
+      expect(declination, true);
+    });
+
+    test('Sun Horizontal position', () {
+      final azimuth = (calc.sunHorizontalPosition.azimuth - 0.06459949466776038).abs() <= _kAnglesAccuracy;
+      final elevation = (calc.sunHorizontalPosition.elevation - -31.246875568260478).abs() <= _kAnglesAccuracy;
+
+      expect(azimuth, true);
+      expect(elevation, true);
+    });
+  });
+
+  final refDate = Instant(2021, 05, 02);
+
+  final date = refDate.add(Duration(hours: 13, minutes: 25, seconds: 35));
 
   Map<String, Map<String, dynamic>> expectedResults = {
     'Barcelona': {
@@ -104,11 +148,9 @@ void main() {
     '80° 12°': {
       'latitude': 80.0,
       'longitude': 12.0,
-      'sunrise': DateTime.utc(2021, 04, 13)
-          .add(Timespan.fromMinutes(-11.654374107739168)),
+      'sunrise': Instant(2021, 04, 13).add(Timespan.fromMinutes(-11.654374107739168)),
       'noon': refDate.add(Duration(hours: 11, minutes: 08)),
-      'sunset': DateTime.utc(2021, 08, 30)
-          .add(Timespan.fromMinutes(1349.7335155238095)),
+      'sunset': Instant(2021, 08, 30).add(Timespan.fromMinutes(1349.7335155238095)),
       'declination': 15.55,
       'azimuth': 216.22,
       'elevation': 23.76,
@@ -116,11 +158,9 @@ void main() {
     '-80° 12°': {
       'latitude': -80.0,
       'longitude': 12.0,
-      'sunrise': DateTime.utc(2021, 08, 25)
-          .add(Timespan.fromMinutes(644.7320304778677)),
+      'sunrise': Instant(2021, 08, 25).add(Timespan.fromMinutes(644.7320304778677)),
       'noon': refDate.add(Duration(hours: 11, minutes: 08)),
-      'sunset': DateTime.utc(2021, 04, 18)
-          .add(Timespan.fromMinutes(678.971344895073)),
+      'sunset': Instant(2021, 04, 18).add(Timespan.fromMinutes(678.971344895073)),
       'declination': 15.55,
       'azimuth': 326.96,
       'elevation': -7.16,
@@ -128,11 +168,9 @@ void main() {
     '90° 12°': {
       'latitude': 90.0,
       'longitude': 12.0,
-      'sunrise': DateTime.utc(2020, 03, 18)
-          .add(Timespan.fromMinutes(370.22985636407043)),
+      'sunrise': Instant(2020, 03, 18).add(Timespan.fromMinutes(370.22985636407043)),
       'noon': refDate.add(Duration(hours: 11, minutes: 08)),
-      'sunset': DateTime.utc(2021, 09, 25)
-          .add(Timespan.fromMinutes(972.5818948340684)),
+      'sunset': Instant(2021, 09, 25).add(Timespan.fromMinutes(972.5818948340684)),
       'declination': 15.56,
       'azimuth': 214.17,
       'elevation': 15.69,
@@ -141,48 +179,34 @@ void main() {
 
   for (var entry in expectedResults.entries) {
     group(entry.key, () {
-      var expected = expectedResults[entry.key]!;
+      final expected = expectedResults[entry.key]!;
 
-      var calc =
-          SolarCalculator(date, expected['latitude'], expected['longitude']);
+      final calc = SolarCalculator(date, expected['latitude'], expected['longitude']);
 
       test('Sunrise', () {
-        var sunrise =
-            calc.sunriseTime.difference(expected['sunrise']).inSeconds.abs() <=
-                _kTimesAccuracyInSec;
+        final sunrise = calc.sunriseTime.difference(expected['sunrise']).inSeconds.abs() <= _kTimesAccuracyInSec;
         expect(sunrise, true);
       });
 
       test('Sunset', () {
-        var sunset =
-            calc.sunsetTime.difference(expected['sunset']).inSeconds.abs() <=
-                _kTimesAccuracyInSec;
+        final sunset = calc.sunsetTime.difference(expected['sunset']).inSeconds.abs() <= _kTimesAccuracyInSec;
         expect(sunset, true);
       });
 
       test('Transit', () {
-        var transit =
-            calc.sunTransitTime.difference(expected['noon']).inSeconds.abs() <=
-                _kTimesAccuracyInSec;
+        final transit = calc.sunTransitTime.difference(expected['noon']).inSeconds.abs() <= _kTimesAccuracyInSec;
         expect(transit, true);
       });
 
       test('Sun Declination', () {
-        var declination =
-            (calc.sunEquatorialPosition.declination - expected['declination'])
-                    .abs() <=
-                _kAnglesAccuracy;
+        final declination =
+            (calc.sunEquatorialPosition.declination - expected['declination']).abs() <= _kAnglesAccuracy;
         expect(declination, true);
       });
 
       test('Sun Horizontal position', () {
-        var azimuth =
-            (calc.sunHorizontalPosition.azimuth - expected['azimuth']).abs() <=
-                _kAnglesAccuracy;
-        var elevation =
-            (calc.sunHorizontalPosition.elevation - expected['elevation'])
-                    .abs() <=
-                _kAnglesAccuracy;
+        final azimuth = (calc.sunHorizontalPosition.azimuth - expected['azimuth']).abs() <= _kAnglesAccuracy;
+        final elevation = (calc.sunHorizontalPosition.elevation - expected['elevation']).abs() <= _kAnglesAccuracy;
 
         expect(azimuth, true);
         expect(elevation, true);