+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.0.1
 
 - Hours of darkness are now calculated based on the civil twilight instead of the nautical twilight.
+
+## 1.0.2
+
+- The instant in time for which to calculate the data is now specified by an Instant object that handle a time zone.
+- Sunrise, sunset, sun transit times and twilight periods are now calculated in the same time zone as the Instant object.